Adding tests (and the required seperate files) for all the PVGroups IOC: NOTE: An issue in caproto with a PR (PR#843) is preventing the tests from running
