+++ tests/test_caproto_servers.py
@@ -14,13 +14,26 @@ except ImportError:
     numpy = None
 
 
-ioc_example_to_info = {"ARI_SXN_SimBeamline.Caproto_servers.Mirror_section":
-                       dict(group_cls='AriM1Mirror',
+ioc_example_to_info = {"ARI_SXN_SimBeamline.caproto_servers.ari_m1":
+                       dict(group_cls='AriM1',
+                            kwargs={},
+                            marks=[pytest.mark.skipif(numpy is None, reason="Requires numpy")],),
+                       "ARI_SXN_SimBeamline.caproto_servers.four_blade_electrometer":
+                       dict(group_cls='FourBladeElectrometer',
+                            kwargs={},
+                            marks=[pytest.mark.skipif(numpy is None, reason="Requires numpy")],),
+                       "ARI_SXN_SimBeamline.caproto_servers.baffle_slit":
+                       dict(group_cls='BaffleSlit',
+                            kwargs={},
+                            marks=[pytest.mark.skipif(numpy is None, reason="Requires numpy")],),
+                       "ARI_SXN_SimBeamline.caproto_servers.diagnostic":
+                       dict(group_cls='Diagnostic',
                             kwargs={},
                             marks=[pytest.mark.skipif(numpy is None, reason="Requires numpy")],)
                        }
 
 
+# noinspection PyTypeChecker
 @pytest.mark.flaky(reruns=2, reruns_delay=2)
 @pytest.mark.parametrize(
     "module_name",
@@ -46,19 +59,19 @@ def test_ioc_examples(request, module_name, async_lib):
 
     skip_pvs = [('ophyd', ':exit')]
 
-    def find_put_value(pv, channeldata):
+    def find_put_value(input_pv, input_channeldata):
         """Determine value to write to pv"""
         for skip_ioc, skip_suffix in skip_pvs:
             if skip_ioc in module_name:
-                if pv.endswith(skip_suffix):
+                if input_pv.endswith(skip_suffix):
                     return None
 
         for put_class, put_value in put_values:
-            if isinstance(channeldata, put_class):
+            if isinstance(input_channeldata, put_class):
                 return put_value
         else:
             raise Exception('Failed to set default value for channeldata:'
-                            f'{channeldata.__class__}')
+                            f'{input_channeldata.__class__}')
 
     for pv, channeldata in info.pvdb.items():
         value = find_put_value(pv, channeldata)
